Explode n.py and move libraries into folder
