+++ python/variable.py
@@ -1,0 +1,4 @@
+class Variable:
+	def __init__(self, t, value):
+		self.type = t
+		self.value = value
